Added a few minor changes to structure of the app
